LayoutTests/media/modern-media-controls/media-controller/ios/media-controller-stop-updates-in-fullscreen-expected.txt

Testing we stop updating controls while in fullscreen and then update them again when back to inline on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Obtained a 'play' event.

Entering fullscreen.

Obtained a webkitpresentationmodechanged event
media.webkitDisplayingFullscreen = true.
PASS media.paused is false

Time progressed.
PASS elapsedTimeWhenEnteringFullscreen === elapsedTime() is true

Obtained a webkitpresentationmodechanged event
media.webkitDisplayingFullscreen = false.
PASS elapsedTimeWhenEnteringFullscreen === elapsedTime() became false

PASS successfullyParsed is true

TEST COMPLETE
